requirements.txt
streamlit==1.56.0
langchain==1.2.15
langchain-core==1.3.0
langchain-text-splitters==1.1.2
langchain-chroma==1.1.0
langchain-community==1.0.0a1
langchain-openai==1.1.15
chromadb==1.5.8
pypdf==6.10.2
python-dotenv==1.2.2